More fixup! added uint32 conversion when creating a new pet. This fixes a bug causing stored pets to be unretrievable.

+++ pkg/pet/memstore.go
@@ -24,7 +24,7 @@ func (m *MemStore) CreatePet(pet *Pet) error {
 	if _, ok := m.store.Load(pet.ID); ok {
 		return Errorf(ErrDuplicate, "Pet with id %d already exists", pet.ID)
 	}
-	m.store.Store(pet.ID, *pet)
+	m.store.Store(uint32(pet.ID), *pet)
 	return nil
 }
